+++ src/components/v2v3/V2V3Project/V2V3ManageTokensSection/LegacyProjectTokensDescription/MigrateLegacyProjectTokensModal/MigrateLegacyProjectTokensModal.tsx
@@ -6,7 +6,7 @@ import { MigrateLegacyProjectTokensForm } from './MigrateLegacyProjectTokensForm
 import { TokenSwapDescription } from './TokenSwapDescription'
 import { useMigrateTokensTx } from 'hooks/JBV3Token/transactor/MigrateTokensTx'
 import { BigNumber } from '@ethersproject/bignumber'
-import { GrantPermissionSection } from './GrantPermissionSection'
+import { GrantApprovalCallout } from './GrantApprovalCallout'
 
 export function MigrateLegacyProjectTokensModal({
   legacyTokenBalance,
@@ -62,7 +62,8 @@ export function MigrateLegacyProjectTokensModal({
     >
       <Space size="large" direction="vertical" className="w-full">
         <TokenSwapDescription />
-        <GrantPermissionSection />
+
+        <GrantApprovalCallout />
 
         <MigrateLegacyProjectTokensForm
           form={form}
